Corrected a bug (with filesystems where the group descriptor table does not fit on a single block).
git-svn-id: svn+ssh://svn.code.sf.net/p/jnode/svn/trunk@447 407389f7-7c16-0410-9eee-ad85acb55d7f

+++ fs/src/fs/org/jnode/fs/ext2/INodeTable.java
@@ -24,7 +24,7 @@ public class INodeTable {
 	public INodeTable(Ext2FileSystem fs, int firstBlock) throws IOException {
 		this.fs = fs;
 		this.firstBlock = firstBlock;
-		blockSize=(int)fs.getBlockSize();
+		blockSize=fs.getBlockSize();
 		blockCount = (int)Math.ceil(
 			(double)(fs.getSuperblock().getINodesPerGroup()*INode.INODE_LENGTH) / 
 			(double) blockSize);
